feat: Gestion des erreurs de l'API

+++ src/routes/deletePokemons.js
@@ -3,8 +3,12 @@ const { Pokemon } = require('../db/sequelize')
 module.exports = (app) => {
   app.delete('/api/pokemons/:id', (req, res) => {
     Pokemon.findByPk(req.params.id).then(pokemon => {
+      if(pokemon === null){
+        const message = `Erreur 404 durant la suppression du pokémon`
+        return res.status(404).json(message)
+      }
       const pokemonDeleted = pokemon;
-      Pokemon.destroy({
+      return Pokemon.destroy({
         where: { id: pokemon.id }
       })
          .then(_ => {
@@ -12,5 +16,9 @@ module.exports = (app) => {
            res.json({message, data: pokemonDeleted })
          })
     })
+       .catch(error =>{
+         const message = `Erreur 500 lors de la suppression du pokémon`
+         return res.status(500).json({message,data: error})
+       })
   })
 }
